Add files via upload
Notebook for data analysis. 
So far it only contains code for uploading of the monthly dataframes for 26 chosen Italian cities for years 1985-2019 included. Dataframes are stored in the csv files packed in the "data.rar" archive. In order to upload them into the Jupyter notebook, download data to your computer and set the appropriate path to them in the code.