Add files via upload
Data scraper. Not object oriented as of now. Downloads data for 52 Italian cities (in two 26-26 parts) for each month in years 1985-2019.
The final output are lists of dataframes where each dataframe contains daily values for a month for a city in a year - 21,840 dataframes in total. These dataframes are then exported to csv files.